chore: add pyright config for editor/LSP type checking
Adds pyrightconfig.json so pyright/pylance give useful output on this repo
instead of ~78 errors that are almost entirely docker-py stub artefacts.

- Targets Python 3.11 to match the python:3.11-slim image
- Excludes tests/ (the dynamic module loader in app_loader.py can't be
  type-checked meaningfully) and limits analysis to app.py
- Downgrades to warning, rather than silencing, the three rule families
  driven by docker-py's incomplete stubs: docker.errors/docker.types not
  being re-exported, Optional-typed Container attributes, and
  APIClient.update_container rejecting a container-id str
- Keeps the checks that would catch real bugs in our own code as errors

Also fixes the annotation on _human_size, which declared n: int and then
immediately rebound it to float. int is assignable to float, so callers
are unaffected.

Net: 78 errors -> 2, both pre-existing and unrelated to this change.
No runtime behaviour changes; the container runs plain `python app.py`
and never reads this config.

+++ app.py
@@ -434,7 +434,7 @@ def _mounts_from_hcfg(hcfg: dict):
     return out or None
 
 
-def _human_size(n: int) -> str:
+def _human_size(n: float) -> str:
     n = float(n or 0)
     for unit in ("B", "KB", "MB", "GB", "TB"):
         if n < 1024 or unit == "TB":
